added long support to math functions

+++ ft_constrain.c
@@ -6,16 +6,16 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/19 12:42:00 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/02 14:01:54 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/09 12:37:36 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_constrain(int n, int num1, int num2)
+long	ft_constrain(long n, long num1, long num2)
 {
-	int	min;
-	int	max;
+	long	min;
+	long	max;
 
 	if (num1 < num2)
 	{
