+++ ft_putnbr_base.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/08 15:28:35 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/04 11:06:18 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/09 12:33:05 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_base(int value, int base)
+void	ft_putnbr_base(long value, int base)
 {
 	ft_putnbr_fd_base(value, 1, base);
 }
