+++ ft_putnbr_fd_base.c
@@ -6,16 +6,16 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/09 12:12:03 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/04 17:42:02 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/09 12:32:47 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_fd_base(int value, int fd, int base)
+void	ft_putnbr_fd_base(long value, int fd, int base)
 {
 	char			*bstr;
-	unsigned int	num;
+	unsigned long	num;
 
 	bstr = "0123456789ABCDEF";
 	num = (value < 0 && base == 10) ? -value : value;
